Generalized Lotka-Volterra Model in Julia

Welcome to the documentation for GLV.jl.

Installation

The package can be installed with

using Pkg; Pkg.add("GLV")

Docstrings

GLV.Community — Type

Community(A, r, K)

Create a community with interaction matrix A, growth rates r, and carrying capacities K.

Example

A = [-1 0; 0 -1]
r = [1, 1]
K = [1, 1]
community = Community(A, r, K)

See also solve.

CommonSolve.solve — Method

DifferentialEquations.solve(c::Community, u0, tspan; kwargs...)

Run the GLV model for community c with initial conditions u0 and time span tspan. The GLV model writes

\[\frac{\mathrm{d} N_i}{\mathrm{d}t} = r_i N_i \left(\frac{\sum_{j\neq i} A_{ij} N_j - N_i}{K_i}\right)\]

where $r$ is the growth rate, $A$ is the interaction matrix, and $K$ is the carrying capacity.

Example

Two non-interacting species with different carrying capacities.

A = [-1 0; 0 -1] # Only self-interactions.
r = [1.0, 1.0]
K = [1.0, 2.0]
c = Community(A, r, K)
u0, tspan = [1.0, 1.0], (0, 10_000) # Simulation parameters.
sol = solve(c, u0, tspan) # Simulate the dynamics.
# At the end of the simulation species should have reached their carrying capacities.
isapprox(sol[end], K, atol=0.01)

See also Community, solve.

GLV.abundance — Method

abundance(c::Community)

Compute the equilibrium abundance of species in community c. Assumes that A is invertible.

Example

The equilibrium abundance of two non-interacting species is equal to their carrying capacities.

A = [-1 0; 0 -1]
r = [1, 1]
K = [1, 2]
c = Community(A, r, K)
abundance(c) == K

See also relative_yield

GLV.relative_yield — Method

relative_yield(c::Community)

Compute the equilibrium relative yield of species in community c. Relative yield is the ratio of abundance to carrying capacity. Assumes that A is invertible.

Example

The equilibrium relative yields of two non-interacting species are equal to the one.

A = [-1 0; 0 -1]
r = [1, 1]
K = [1, 2]
c = Community(A, r, K)
relative_yield(c) == [1, 1]

See also abundance.