Generalized Lotka-Volterra Model in Julia

Welcome to the documentation for GLV.jl.

Installation

The package can be installed with

using Pkg; Pkg.add("GLV")

Docstrings

GLV.Community — Type

Community(A, r, K)

Create a community with interaction matrix A, growth rates r, and carrying capacities K.

Example

A = [-1 0; 0 -1]
r = [1, 1]
K = [1, 1]
community = Community(A, r, K)

See also solve, simulate_pulse.

Base.rand — Method

Base.rand(
::Type{Community},
S::Int;
A_ij::Distribution=Normal(0, 1),
r_i::Union{Nothing,Distribution}=nothing,
K_i::Union{Nothing,Distribution}=nothing,

)

Genereate a random community with S species. Parameters are drawn from specified distributions. By default, species growth rates and carrying capacities are set to one. Species self-regulation, that is the diagonal of A, is set to -1.

Example

using Distributions
c = rand(Community, 10; A_ij=Normal(-1, 0.1))

See also Community.

CommonSolve.solve — Method

DifferentialEquations.solve(c::Community, u0, tspan; kwargs...)

Run the GLV model for community c with initial conditions u0 and time span tspan. The GLV model writes

\[\frac{\mathrm{d} N_i}{\mathrm{d}t} = r_i N_i \left(\frac{\sum_{j\neq i} A_{ij} N_j - N_i}{K_i}\right)\]

where $r$ is the growth rate, $A$ is the interaction matrix, and $K$ is the carrying capacity.

Example

Two non-interacting species with different carrying capacities.

A = [-1 0; 0 -1] # Only self-interactions.
r = [1.0, 1.0]
K = [1.0, 2.0]
c = Community(A, r, K)
u0, tspan = [1.0, 1.0], (0, 10_000) # Simulation parameters.
sol = solve(c, u0, tspan) # Simulate the dynamics.

See also Community.

GLV.abundance — Method

abundance(c::Community)

Compute the equilibrium abundance of species in community c. Assumes that A is invertible.

Example

The equilibrium abundance of two non-interacting species is equal to their carrying capacities.

A = [-1 0; 0 -1]
r = [1, 1]
K = [1, 2]
c = Community(A, r, K)
abundance(c) == K

See also relative_yield

GLV.offdiag — Method

offdiag(A)

Dictionnary of off-diagonal elements of a matrix A. Keys are pairs of indices (i, j) and values are the corresponding elements of A.

A = [1 2; 3 4]
offdiag(A)

GLV.relative_yield — Method

relative_yield(c::Community)

Compute the equilibrium relative yield of species in community c. Relative yield is the ratio of abundance to carrying capacity. Assumes that A is invertible.

Example

The equilibrium relative yields of two non-interacting species are equal to the one.

A = [-1 0; 0 -1]
r = [1, 1]
K = [1, 2]
c = Community(A, r, K)
relative_yield(c) == [1, 1]

See also abundance.

GLV.simulate_extinctions — Method

simulate_extinctions(c::Community, idx, tspan)

Simulate the dynamics of the community c after the extinction of the species of indices idx for the time span tspan.

Example

c = rand(Community, 5; A_ij=Normal(0, 0.1))
simulate_extinctions(c, [1, 3], (0, 100)) # Species 1 and 3 go extinct.

See also solve, simulate_pulse.

GLV.simulate_pulse — Method

simulate_pulse(c::Community, x, tspan)

Simulate the recovery of the community c after the pulse perturbation x for the time span tspan. In other words, it simulate the community dynamics of initial conditions $N^* + x$ where N^* is the vector of species equilibrium abundances.

Example

using Distributions
S = 5 # Number of species.
c = rand(Community, S; A_ij=Normal(0, 0.1))
x = rand(Normal(-2, 0.1), S)
u = simulate_pulse(c, x, (0, 100))

See also solve.